+++ client/typescript-sdk/build.sh
@@ -1,0 +1,3 @@
+#!/bin/bash -e
+yarn install --production=false
+yarn run build
