+++ client/typescript-sdk/src/index.ts
@@ -1,0 +1,2 @@
+export * from "./orbs-client";
+export * from "./public-api-client";
