+++ client/typescript-sdk/src/public-api-client.ts
@@ -1,0 +1,18 @@
+import * as path from "path";
+import * as caller from "grpc-caller";
+import { PublicApiClient, getPathToProto } from "orbs-interfaces";
+
+export interface GRPCClient {
+  proto: string;
+  name: string;
+  endpoint: string;
+}
+
+function client(grpc: GRPCClient) {
+  const protoPath = getPathToProto(grpc.proto);
+  return caller(grpc.endpoint, protoPath, grpc.name);
+}
+
+export function initPublicApiClient({ endpoint }: { endpoint: string }): PublicApiClient {
+  return client({ proto: "public-api.proto", name: "PublicApi", endpoint });
+}
